Flavour text and Icon update
Added my missing flavour text at the start of the game in the form of a welcome screen, and added a button that says "play" to start the quiz. Updated the score screen at the end of the game to add flavour text. I noticed that a certain icon appeared at the score screen, and I wanted to update it! Updating this icon also updated the files icon, so two birds in one stone!